+++ src/main/java/com/example/lxp/review/application/port/in/dto/DeleteReviewsByCourseCommand.java
@@ -1,0 +1,8 @@
+package com.example.lxp.review.application.port.in.dto;
+
+import jakarta.validation.constraints.NotNull;
+
+public record DeleteReviewsByCourseCommand(
+        @NotNull Long courseId
+) {
+}
